moved enclave test runner into its own package
