+++ cosmwasm/Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 members = ["packages/*"]
-exclude = ["contracts", "packages/wasmi-runtime"]
+exclude = ["contracts", "packages/wasmi-runtime", "packages/enclave-test"]
