+++ cosmwasm/packages/enclave-test/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "enclave-test"
+version = "0.1.0"
+authors = ["Reuven Podmazo <reuven@enigma.co>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+cosmwasm-sgx-vm = { path = "../sgx-vm", features = ["iterator", "enclave-tests"] }
