--- go-cosmwasm/src/enclave_tests.rs
@@ -1,1 +1,0 @@
-pub use cosmwasm_sgx_vm::enclave_tests::*;
